Updated Interaction System
